fix: share event cursor policy with app tests
